cv_service/requirements.txt
torch
torchvision
opencv-python
ultralytics
kafka-python
numpy